-아티클 썸네일, 삭제 기능

+++ controllers/articleController.js
@@ -67,18 +67,18 @@ exports.view = async (req, res) => {
         }
 
         let interaction = { liked: false, bookmarked: false };
-        let user = { role: null, id: null };
+        let userData = { role: null, id: null };
         if (req.session.user) {
-            user.role = req.session.user.role;
-            user.id = req.session.user.id;
+            userData.role = req.session.user.role;
+            userData.id = req.session.user.id;
             
-            if (user.role == 'patient') {
-                if (await ArticleInteractionModel.findLikeByPatientAndArticle(articleId, user.id) != null) interaction.liked = true;
-                if (await ArticleInteractionModel.findBookmarkByPatientAndArticle(articleId, user.id) != null) interaction.bookmarked = true;
+            if (userData.role == 'patient') {
+                if (await ArticleInteractionModel.findLikeByPatientAndArticle(articleId, userData.id) != null) interaction.liked = true;
+                if (await ArticleInteractionModel.findBookmarkByPatientAndArticle(articleId, userData.id) != null) interaction.bookmarked = true;
             }
         }      
 
-        res.render('article/view', { user, article, counselor, htmlContent, interaction });
+        res.render('article/view', { userData, article, counselor, htmlContent, interaction });
     } catch (error) {
         console.error("viewDiary 오류:", error);
         res.status(500).send("서버 오류가 발생했습니다.");
@@ -86,49 +86,48 @@ exports.view = async (req, res) => {
 }
 
 
-// exports.delete = async (req, res) => {
-//     try {
-//         if (req.session.user) {
-//             const articleId = req.params.articleId;
-//             const article = await ArticleModel.findById(articleId); 
-//             const counselor = await UserModel.getCounselorById(article.counselor_id);
+exports.delete = async (req, res) => {
+    try {
+        if (req.session.user) {
+            const articleId = req.params.articleId;
+            const article = await ArticleModel.findById(articleId); 
             
-//             if (!AccessCheck.checkPatientId(req.session.user.role, req.session.user.id, diary.patient_id)) {
-//                 const referer = req.get('Referer') || '/';
-//                 return res.status(403).send(`<script>alert("권한이 없습니다."); window.location.href = "${referer}";</script>`);
-//             } 
+            if (!AccessCheck.checkCounselorId(req.session.user.role, req.session.user.id, article.counselor_id)) {
+                const referer = req.get('Referer') || '/';
+                return res.status(403).send(`<script>alert("권한이 없습니다."); window.location.href = "${referer}";</script>`);
+            } 
     
-//             // 스토리지 이미지 삭제
-//             if (diary.image_url) {
-//                 // URL parsing
-//                 const parsedUrl = new URL(diary.image_url);
-//                 const encodedFilePath = parsedUrl.pathname.split('/').pop();
-//                 const filePath = decodeURIComponent(encodedFilePath);
+            // 스토리지 내용 삭제
+            if (article.content) {
+                // URL parsing
+                const parsedUrl = new URL(article.content);
+                const encodedFilePath = parsedUrl.pathname.split('/').pop();
+                const filePath = decodeURIComponent(encodedFilePath);
     
-//                 const admin = require('firebase-admin');
-//                 const bucket = admin.storage().bucket('comma-5a85c.appspot.com'); 
-//                 const file = bucket.file(filePath);
+                const admin = require('firebase-admin');
+                const bucket = admin.storage().bucket('comma-5a85c.appspot.com'); 
+                const file = bucket.file(filePath);
     
-//                 try {
-//                     await file.delete();  // 파일 삭제
-//                 } catch (error) {
-//                     console.error("Failed to delete file", error);
-//                     res.status(500).send({ error: "Failed to delete file: " + error.message });
-//                 }
-//             }
+                try {
+                    await file.delete();  // 파일 삭제
+                } catch (error) {
+                    console.error("Failed to delete file", error);
+                    res.status(500).send({ error: "Failed to delete file: " + error.message });
+                }
+            }
             
-//             // DB diary 삭제
-//             await diaryModel.delete(diaryId);
+            // DB diary 삭제
+            await ArticleModel.delete(articleId);
     
-//             // redirect
-//             return res.json({ success: true, redirect: `/profile/patient/${patient.id}/diaries` });
-//         }
-//         else return res.render("login/login");
-//     } catch (error) {
-//         console.error("deleteDiary 오류:", error);
-//         res.status(500).send("서버 오류가 발생했습니다.");
-//     }
-// }
+            // redirect
+            return res.json({ success: true, redirect: `/article` });
+        }
+        else return res.render("login/login");
+    } catch (error) {
+        console.error("deleteDiary 오류:", error);
+        res.status(500).send("서버 오류가 발생했습니다.");
+    }
+}
 
 // 상담사 메인화면 일기 리스트
 exports.list = async (req, res) => {
@@ -141,8 +140,7 @@ exports.list = async (req, res) => {
 
         if (Previews) {
             Previews.forEach(preview => {
-                preview.image_url = setDefaultImage(preview.image_url);
-                // preview.profile_picture = 프로필 기본이미지 설정
+                preview.thumbnail_url = setDefaultImage(preview.thumbnail_url);
             });
         }
         
@@ -158,21 +156,28 @@ exports.list = async (req, res) => {
 exports.toggleLike = async (req, res) => { 
     try {
         // 로그인 여부 확인 && 요청id == 세션id
+        if (req.session.user) {
+            if (!AccessCheck.checkPatientRole(req.session.user.role)) {
+                const referer = req.get('Referer') || '/';
+                return res.status(403).send(`<script>alert("권한이 없습니다."); window.location.href = "${referer}";</script>`);
+            }
 
-        let { articleId, patientId } = req.body;
-        articleId = parseInt(articleId); // 추출한 body값을 정수로 변환
-        patientId = parseInt(patientId);
-
-        let liked = true;
-        if (await ArticleInteractionModel.findLikeByPatientAndArticle(articleId, patientId) == null) { // 좋아요 추가
-            ArticleInteractionModel.createLike(articleId, patientId);
-        }
-        else { // 좋아요 삭제
-            ArticleInteractionModel.deleteLike(articleId, patientId);
-            liked = false;
-        }
+            let { articleId, patientId } = req.body;
+            articleId = parseInt(articleId); // 추출한 body값을 정수로 변환
+            patientId = parseInt(patientId);
     
-        return res.json({ response: true, liked });
+            let liked = true;
+            if (await ArticleInteractionModel.findLikeByPatientAndArticle(articleId, patientId) == null) { // 좋아요 추가
+                ArticleInteractionModel.createLike(articleId, patientId);
+            }
+            else { // 좋아요 삭제
+                ArticleInteractionModel.deleteLike(articleId, patientId);
+                liked = false;
+            }
+        
+            return res.json({ response: true, liked });
+        }
+        else return res.render("login/login");
     } catch (error) {
         console.error("좋아요 오류:", error);
     }
@@ -182,21 +187,28 @@ exports.toggleLike = async (req, res) => {
 exports.toggleBookmark = async (req, res) => { 
     try {
         // 로그인 여부 확인 && 요청id == 세션id
+        if (req.session.user) {
+            if (!AccessCheck.checkPatientRole(req.session.user.role)) {
+                const referer = req.get('Referer') || '/';
+                return res.status(403).send(`<script>alert("권한이 없습니다."); window.location.href = "${referer}";</script>`);
+            }
 
-        let { articleId, patientId } = req.body;
-        articleId = parseInt(articleId); // 추출한 body값을 정수로 변환
-        patientId = parseInt(patientId);
+            let { articleId, patientId } = req.body;
+            articleId = parseInt(articleId); // 추출한 body값을 정수로 변환
+            patientId = parseInt(patientId);
 
-        let bookmarked = true;
-        if (await ArticleInteractionModel.findBookmarkByPatientAndArticle(articleId, patientId) == null) { // 북마크 추가
-            ArticleInteractionModel.createBookmark(articleId, patientId);
+            let bookmarked = true;
+            if (await ArticleInteractionModel.findBookmarkByPatientAndArticle(articleId, patientId) == null) { // 북마크 추가
+                ArticleInteractionModel.createBookmark(articleId, patientId);
+            }
+            else { // 북마크 삭제
+                ArticleInteractionModel.deleteBookmark(articleId, patientId);
+                bookmarked = false;
+            }
+        
+            return res.json({ response: true, bookmarked });
         }
-        else { // 북마크 삭제
-            ArticleInteractionModel.deleteBookmark(articleId, patientId);
-            bookmarked = false;
-        }
-    
-        return res.json({ response: true, bookmarked });
+        else return res.render("login/login");
     } catch (error) {
         console.error("북마크 오류:", error);
     }
